Use tzdb with `std` feature
Shadow-rs uses `std::error::Error`s, which are only implemented by
tz-rs, and tzdb if their features `std` are enabled. Right now this is
not a problem, because even if you are using tzdb without default
features, tzdb still uses tz-rs with default-features. This is because
tz-rs did not yet have the feature gate when tzdb v0.3 was released.

Once tzdb v0.4 is released, the unconditional use of tzdb's std feature
will be fixed, which would break this project. So this PR is proactively
fixing the issue.